Add Logic folder and move calculate and operate modules
