+++ src/components/App.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import Display from './Display';
 import ButtonPanel from './ButtonPanel';
-import Calculate from '../calculate';
+import Calculate from '../logic/calculate';
 
 function App() {
   return (
